refactor: Reduce boilerplate with generic TransactionablePassthrough protocol
