+++ MailCore/Cache/MailboxInfosManager/MailboxInfosManager+Transactionable.swift
@@ -17,27 +17,6 @@
  */
 
 import Foundation
-import InfomaniakCoreDB
-import RealmSwift
 
 /// Transactionable
-extension MailboxInfosManager: Transactionable {
-    public func fetchObject<Element: Object, KeyType>(ofType type: Element.Type,
-                                                      forPrimaryKey key: KeyType) -> Element? {
-        return transactionExecutor.fetchObject(ofType: type, forPrimaryKey: key)
-    }
-
-    public func fetchObject<Element: RealmFetchable>(ofType type: Element.Type,
-                                                     filtering: (Results<Element>) -> Element?) -> Element? {
-        return transactionExecutor.fetchObject(ofType: type, filtering: filtering)
-    }
-
-    public func fetchResults<Element: RealmFetchable>(ofType type: Element.Type,
-                                                      filtering: (Results<Element>) -> Results<Element>) -> Results<Element> {
-        return transactionExecutor.fetchResults(ofType: type, filtering: filtering)
-    }
-
-    public func writeTransaction(withRealm realmClosure: (Realm) throws -> Void) throws {
-        try transactionExecutor.writeTransaction(withRealm: realmClosure)
-    }
-}
+extension MailboxInfosManager: TransactionablePassthrough {}
